fix: router bug in activitypub router
We had a middleware applied in the activitypub router which would return 404 for all routes if activitypub wasn't enabled. Unfortunately due to the way Express works, that middleware also applied to every route below that router's routes in the base router chain. Instead, we now apply the middleware specifically to the individual routes in the activitypub file.